Feat: added authentication to the API functionalities

+++ pages/api/products/[id].js
@@ -5,7 +5,10 @@ export default async function handler(req, res) {
   const {
     method,
     query: { id },
+    cookies,
   } = req;
+
+  const token = cookies.token;
 
   dbConnect();
 
@@ -18,6 +21,9 @@ export default async function handler(req, res) {
     }
   }
   if (method === "PUT") {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json("Not authenticated!");
+    }
     try {
       const product = await Product.create(req.body);
       res.status(201).json(product);
@@ -26,6 +32,9 @@ export default async function handler(req, res) {
     }
   }
   if (method === "DELETE") {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json("Not authenticated!");
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json("Product has been deleted!");
